+++ CHANGELOG.md
@@ -50,6 +50,7 @@
   - 更新 `README.md` 中的链接指向。
 
 - 实现运维指令。
+  - com.dwarfeng.fdr.impl.service.telqos.SourceCommand。
   - com.dwarfeng.fdr.impl.service.telqos.SupportCommand。
 
 - 优化支持实体机制。